requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Config
python-dotenv==1.0.1
pydantic-settings==2.5.2

# ML
scikit-learn==1.5.2
xgboost==2.1.1
pandas==2.2.3
numpy==1.26.4
joblib==1.4.2

# Forecasting
prophet==1.1.5

# RAG / LLM
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
langgraph==0.2.28
faiss-cpu==1.8.0
sentence-transformers==3.1.1
openai==1.51.0

# Utilities
httpx==0.27.2
